Add files via upload
Code used to create groups of interest ('CN A-', 'CN A+ Hyp-', 'CN A+ Hyp+', 'MCI A+ Hyp-', 'MCI A+ Hyp+', 'AD A+ Hyp-', 'AD A+ Hyp+'). ANCOVA analysis was performed mainly for 'CN A+ Hyp- vs CN A+ Hyp-', 'MCI A+ Hyp+ vs MCI A+ Hyp-' and 'AD A+ Hyp+ vs AD A+ Hyp-'.